+++ index.html
@@ -232,6 +232,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script type="text/javascript" src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
