+++ js/data.js
@@ -1,0 +1,30 @@
+const MAX_STRING_LENGTH = 140;
+const PHOTO_DESC = [
+  'Хорошая фотография!',
+  'Блин, а горизонт то завален',
+  'Вы такие счастливые тут',
+  'Это где?'
+];
+const COMMENT_MESS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const COMMENT_NAME = [
+  'Даниил Шаров',
+  'Илья Марков',
+  'Альбина Тетерина',
+  'Аврор Потапов',
+  'Леонид Трифонов',
+  'Лада Дмитриева'
+];
+
+const createPublicationIdArray = [];
+const createPublicationUrlArray = [];
+const createPublicationCommentIdArray = [];
+
+export {MAX_STRING_LENGTH, PHOTO_DESC, COMMENT_MESS, COMMENT_NAME, createPublicationIdArray, createPublicationUrlArray, createPublicationCommentIdArray};
